+++ public/js/components/schemadetail/SchemaDetail.js
@@ -2,8 +2,8 @@ import format_size from "../../util/format_size.js";
 import format_time from "../../util/format_time.js";
 
 import ModalPrompt from "../ModalPrompt.js";
-import ClipboardCopy from "../ClipboardCopy.js";
 import SchemaStar from "./SchemaStar.js";
+import SchemaDownload from "./SchemaDownload.js";
 
 import {
     schema_update,
@@ -16,13 +16,12 @@ import {
 import { get_tags } from "../../service/tags.js";
 import { is_logged_in, has_permission } from "../../service/login.js";
 
-const max_placement_tool_size = 100;
 
 export default {
     components: {
         "modal-prompt": ModalPrompt,
-        "clipboard-copy": ClipboardCopy,
-        "schema-star": SchemaStar
+        "schema-star": SchemaStar,
+        "schema-download": SchemaDownload
     },
     props: {
         search_result: { type: Object, required: true },
@@ -50,7 +49,6 @@ export default {
         get_tags: function() {
             return get_tags().filter(t => !t.restricted || has_permission("ADMIN"));
         },
-
         save: function() {
             this.error_response = null;
             schema_update(this.schema)
@@ -97,12 +95,7 @@ export default {
         }
     },
     computed: {
-        logged_in: is_logged_in,
-        can_use_placement_tool: function() {
-            return (this.schema.size_x <= max_placement_tool_size &&
-                    this.schema.size_y <= max_placement_tool_size &&
-                    this.schema.size_z <= max_placement_tool_size);
-        }
+        logged_in: is_logged_in
     },
     template: /*html*/`
     <div>
@@ -274,84 +267,7 @@ export default {
                         Download
                     </div>
                     <div class="card-body">
-                        <div class="row">
-                            <div class="col-md-6">
-                                <u>
-                                    <h4>Online</h4>
-                                </u>
-                                <h5>Single placement</h5>
-                                <ul>
-                                    <li>
-                                        Download and install the <router-link to="/mod">blockexchange</router-link>-mod
-                                    </li>
-                                    <li>
-                                        Select the origin position with <clipboard-copy :text="'/bx_pos1'"></clipboard-copy>
-                                    </li>
-                                    <li>
-                                        Check if the placement fits with <clipboard-copy :text="'/bx_allocate ' + username + ' ' + schema.name"></clipboard-copy>
-                                    </li>
-                                    <li>
-                                        Place the schematic with <clipboard-copy :text="'/bx_load ' + username + ' ' + schema.name"></clipboard-copy>
-                                    </li>
-                                </ul>
-                                <h5 v-if="can_use_placement_tool">Multiple placements with the placement tool</h5>
-                                <ul v-if="can_use_placement_tool">
-                                    <li>
-                                        Download and install the <router-link to="/mod">blockexchange</router-link>-mod
-                                    </li>
-                                    <li>
-                                        Create a placement tool with <clipboard-copy :text="'/bx_placer ' + username + ' ' + schema.name"></clipboard-copy>
-                                    </li>
-                                    <li>
-                                        Point and click to place the schematic
-                                    </li>
-                                </ul>
-                            </div>
-                            <div class="col-md-6">
-                                <u>
-                                    <h4>Offline</h4>
-                                </u>
-                                <div class="btn-group">
-                                    <a class="btn btn-outline-success"
-                                        :href="BaseURL + '/api/export_bx/' + schema.uid + '/' + schema.name + '.zip'">
-                                        <i class="fa fa-download"></i>
-                                        Download as BX schematic
-                                    </a>
-                                    <a class="btn btn-outline-success"
-                                        v-bind:class="{disabled: schema.total_parts >= 200}"
-                                        :href="BaseURL + '/api/export_we/' + schema.uid + '/' + schema.name + '.we'">
-                                        <i class="fa fa-download"></i>
-                                        Download as WE schematic
-                                        <i class="fa fa-triangle-exclamation"
-                                            v-if="schema.total_parts >= 200"
-                                            style="color: red;"
-                                            title="WE-Import disable due to schematic size"></i>
-                                        <i class="fa fa-triangle-exclamation"
-                                            v-else-if="schema.total_parts > 50"
-                                            style="color: yellow;"
-                                            title="WE-Import might be slow due to schematic size"></i>
-                                    </a>
-                                </div>
-                                <ul>
-                                    <li>
-                                        Download and install the <router-link to="/mod">blockexchange</router-link>-mod
-                                    </li>
-                                    <li>
-                                        Download the Blockexchange schematic with the above button and place it in the
-                                        <b class="text-muted">{world-folder}/bxschems/</b> folder
-                                    </li>
-                                    <li>
-                                        Select the origin position with <clipboard-copy :text="'/bx_pos1'"></clipboard-copy>
-                                    </li>
-                                    <li>
-                                        Check if the placement fits with <clipboard-copy :text="'/bx_allocate_local ' + username + ' ' + schema.name"></clipboard-copy>
-                                    </li>
-                                    <li>
-                                        Place the schematic with <clipboard-copy :text="'/bx_load_local ' + username + ' ' + schema.name"></clipboard-copy>
-                                    </li>
-                                </ul>
-                            </div>
-                        </div>
+                        <schema-download :schema="schema" :username="username"/>
                     </div>
                 </div>
             </div>
